add select categories

+++ pkg/model/chat_category.go
@@ -1,9 +1,9 @@
 package model
 
 type ChatCategory struct {
-	ChatId       int    `db:"chat_id"`
+	ChatId       int64  `db:"chat_id"`
 	ChatName     string `db:"chat_name"`
-	CategoryId   int    `db:"category_id"`
+	CategoryId   int8   `db:"category_id"`
 	CategoryCode string `db:"category_code"`
 	CategoryName string `db:"category_name"`
 }
